app/requirements.txt
fastapi==0.95.1
jinja2==3.1.2
numpy==1.21.0
pandas==1.3.5
python-dotenv==0.19.0
urllib3==1.26.9
uvicorn==0.20.0
websockets==11.0.2
websocket-client==1.5.1
pydantic==1.10.2
SQLAlchemy==1.4.41
mysqlclient==2.1.1
APScheduler==3.10.1
requests==2.28.2
